refactorig check method name and moving it to assertion class
