+++ src/test/java/com/wikia/webdriver/common/core/url/UrlChecker.java
@@ -21,20 +21,6 @@ public class UrlChecker {
     );
   }
 
-  /**
-   * This method checks that URL equals current URL and logs result
-   * The method is case-insensitive 
-   */
-  public static void assertUrlEqualToCurrentUrl(WebDriver driver, String url) {
-    String currentUrl = driver.getCurrentUrl().toLowerCase();
-    PageObjectLogging.log(
-        "Log Url",
-        "Url " + url + " is equal to current Url " + currentUrl,
-        "Url " + url + " isn't equal to current Url " + currentUrl,
-        isUrlEqualToCurrentUrl(driver, url)
-     );
-  }
-
   public static boolean isUrlEqualToCurrentUrl(WebDriver driver, String url) {
     return driver.getCurrentUrl().equalsIgnoreCase(url);
   }
