+++ src/test/java/com/wikia/webdriver/testcases/mercurytests/RedirectsTests.java
@@ -4,9 +4,9 @@ import com.wikia.webdriver.common.contentpatterns.MercurySubpages;
 import com.wikia.webdriver.common.core.Assertion;
 import com.wikia.webdriver.common.core.annotations.Execute;
 import com.wikia.webdriver.common.core.configuration.Configuration;
-import com.wikia.webdriver.common.core.url.UrlChecker;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.ArticlePageObject;
+
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
@@ -19,7 +19,7 @@ public class RedirectsTests extends NewTestTemplate {
 
   @BeforeMethod(alwaysRun = true)
   public void prepareURL() {
-    String articleUrl = urlBuilder.getUrlForPathWithWWW(Configuration.getWikiName(), ARTICLE_NAME);
+    String articleUrl = urlBuilder.getWebUrlForPath(Configuration.getWikiName(), ARTICLE_NAME);
     url = urlBuilder.appendQueryStringToURL(articleUrl, QUERY_STRING);
 
     String expectedArticleUrl = urlBuilder.getUrlForPath(Configuration.getWikiName(), ARTICLE_NAME);
@@ -31,9 +31,6 @@ public class RedirectsTests extends NewTestTemplate {
   public void RedirectsTest_001_RedirectFromWWW() {
     ArticlePageObject article = new ArticlePageObject(driver);
     article.openWikiPage(url);
-    Assertion.assertTrue(
-      UrlChecker.isUrlEqualToCurrentUrl(driver, expectedUrl),
-      String.format("URL is different than expected. Expected: %s \n", expectedUrl)
-    );
+    Assertion.assertUrlEqualToCurrentUrl(driver, expectedUrl);
   }
 }
