+++ src/test/java/com/wikia/webdriver/testcases/mercurytests/curatedcontenttests/MainPageTests.java
@@ -3,8 +3,8 @@ package com.wikia.webdriver.testcases.mercurytests.curatedcontenttests;
 import com.wikia.webdriver.common.contentpatterns.MercuryMessages;
 import com.wikia.webdriver.common.contentpatterns.MercurySubpages;
 import com.wikia.webdriver.common.contentpatterns.MercuryWikis;
+import com.wikia.webdriver.common.core.Assertion;
 import com.wikia.webdriver.common.core.annotations.RelatedIssue;
-import com.wikia.webdriver.common.core.url.UrlChecker;
 import com.wikia.webdriver.common.logging.PageObjectLogging;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.curatedcontent.CuratedContentPageObject;
@@ -464,6 +464,6 @@ public class MainPageTests extends NewTestTemplate {
         .clickOnCuratedContentElementByIndex(2)
         .isAlertNotificationVisible()
         .waitForLoadingSpinnerToFinish();
-    UrlChecker.assertUrlEqualToCurrentUrl(driver, oldUrl);
+    Assertion.assertUrlEqualToCurrentUrl(driver, oldUrl);
   }
 }
