+++ src/test/java/com/wikia/webdriver/testcases/mercurytests/curatedcontenttests/NavigationTests.java
@@ -3,6 +3,7 @@ package com.wikia.webdriver.testcases.mercurytests.curatedcontenttests;
 import com.wikia.webdriver.common.contentpatterns.MercuryPaths;
 import com.wikia.webdriver.common.contentpatterns.MercurySubpages;
 import com.wikia.webdriver.common.contentpatterns.MercuryWikis;
+import com.wikia.webdriver.common.core.Assertion;
 import com.wikia.webdriver.common.core.url.UrlChecker;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.mercury.curatedcontent.CuratedContentPageObject;
@@ -49,10 +50,10 @@ public class NavigationTests extends NewTestTemplate {
     UrlChecker.isPathContainedInCurrentUrl(driver, MercuryPaths.ROOT_PATH);
 
     driver.navigate().back();
-    UrlChecker.assertUrlEqualToCurrentUrl(driver, previousUrl);
+    Assertion.assertUrlEqualToCurrentUrl(driver, previousUrl);
 
     driver.navigate().forward();
-    UrlChecker.assertUrlEqualToCurrentUrl(driver, nextUrl);
+    Assertion.assertUrlEqualToCurrentUrl(driver, nextUrl);
   }
 
   // CCT07
@@ -113,17 +114,17 @@ public class NavigationTests extends NewTestTemplate {
     CuratedContentPageObject section = new CuratedContentPageObject(driver);
     String expectedUrl = wikiURL + MercurySubpages.CC_CATEGORY_TEMPLATES;
     section.navigateToUrlWithPath(wikiURL, MercurySubpages.CC_CATEGORY_TEMPLATES);
-    UrlChecker.assertUrlEqualToCurrentUrl(driver, expectedUrl);
+    Assertion.assertUrlEqualToCurrentUrl(driver, expectedUrl);
 
     expectedUrl = wikiURL + MercurySubpages.CC_SECTION_CATEGORIES;
     section.navigateToUrlWithPath(wikiURL, MercurySubpages.CC_SECTION_CATEGORIES);
-    UrlChecker.assertUrlEqualToCurrentUrl(driver, expectedUrl);
+    Assertion.assertUrlEqualToCurrentUrl(driver, expectedUrl);
 
     expectedUrl = wikiURL + "wiki/mercury_cc_wikia";
     section.navigateToUrlWithPath(wikiURL, MercurySubpages.CC_EMPTY_CATEGORY);
-    UrlChecker.assertUrlEqualToCurrentUrl(driver, expectedUrl);
+    Assertion.assertUrlEqualToCurrentUrl(driver, expectedUrl);
 
     section.navigateToUrlWithPath(wikiURL, MercurySubpages.CC_NOT_EXISTING_SECTION);
-    UrlChecker.assertUrlEqualToCurrentUrl(driver, expectedUrl);
+    Assertion.assertUrlEqualToCurrentUrl(driver, expectedUrl);
   }
 }
